requirements.txt
python-bithumb
openai
streamlit
pandas
plotly
python-dotenv
requests
